requirements.txt
ffmpeg-python>=0.2.0
numpy>=1.16, <=1.23.5
torch>=2.5.1
torchvision>=0.20.1
tqdm>=4.35.0
opencv-python>=4.1.2
cupy-cuda11x